auth added

+++ blog-server-api/src/endpoints/delete_post/handler.rs
@@ -4,7 +4,11 @@ use super::response_content_failure::DeletePostResponseContentFailure::*;
 use super::response_content_success::DeletePostResponseContentSuccess;
 
 pub async fn http_handler(
-    (DeletePostRequestContent { id, post_service },): (DeletePostRequestContent,),
+    (DeletePostRequestContent {
+        id,
+        post_service,
+        auth_author_future,
+    },): (DeletePostRequestContent,),
 ) -> Result<DeletePostResponseContentSuccess, DeletePostResponseContentFailure> {
     let id = id.parse::<u64>().map_err(|e| IncorrectIdFormat {
         reason: e.to_string(),
@@ -16,13 +20,21 @@ pub async fn http_handler(
         });
     }
 
-    let _ = post_service
+    let author = auth_author_future.await.map_err(|e| Unauthorized {
+        reason: e.to_string(),
+    })?;
+
+    let post = post_service
         .post_by_id(&id)
         .await
         .map_err(|e| DatabaseError {
             reason: e.to_string(),
         })?
         .ok_or(NotFound)?;
+
+    if author.id != post.base.author_id {
+        return Err(EditingForbidden);
+    }
 
     post_service
         .delete_post_by_id(&id)
